+++ src/scripts/utilities/tipsTricks.ts
@@ -1,0 +1,22 @@
+
+/**
+ * Toggles dev tools in the current window
+ */
+export const tipsTricks: string[] = [
+  "There is a whole list of really helpful keybind that can make your use of FA a lot quicker and easier!",
+  "The settings menu contains a whole assortment of both big and small tweaks to tailor the app to your needs!",
+  "FA has a Dark Mode that is no mere afterthought and is fully serviceable!",
+  "Many parts of the app react to a middle-mouse click. For example, you can close document tabs or open hierarchical tree items with it directly!",
+  "Any field in the document editing mode with the small dropdown arrow that is not a Relationship field allows for custom values! Just close the menu of predefined values with ESC, type your desired value and confirm with ENTER.",
+  "Fantasia Archive is for free, no string attached! If you ever want to support the project, drop me a penny or two on Patreon!",
+  "This software is actually a local website pretend to be a desktop application! Technology is crazy!",
+  "If you ever need a good, offline map-making software, look at Wonderdraft and Dungeondraft by Megasploot!",
+  "If these tips annoy you, there is a switch to turn them off in the program settings.",
+  "The Relationship searching functionality is packed to the brim with extra features you can use to find what you need! Read all about it in the \"Advanced search guide\" in the Help menu!",
+  "Your project gets auto-saved on exit, but it is suggested to use the \"Export\" function once in a while to make sure the data stays safe.",
+  "The left hierarchical tree is expandable with a mouse-drag and the program remembers where you left it the next time you launch!",
+  "Don't hesitate to add as many sub-categories as you want. You can always move stuff around later and reorganize your project structure.",
+  "Tags are a very powerful feature that allows for adding of semi-categories to the hierarchical tree and additional search parameter for quick navigation of important bits of your story!",
+  "There is a roadmap of planned features on the Patreon! If you are curious what is eventually coming, feel free to read it.",
+  "DISCLAIMER: No magical creatures were harmed during the making of this software."
+]